+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Currently under development
 ---------------------------
 
 - Bug: It was not possible to compile and use external modules
+- Bug: Relative paths did not work in a configuration file, when it was called from another directory
 - Enhancement: Added a command to allow listing of all loaded modules `uipf --list`
 - Enhancement: Added Default and Copy Constructors to the Data types